chore: add clickhouse issue - database is not resolved in the CTE with join

+++ databases/test/migrations/20220415224306_create_sales_table.sql
@@ -1,9 +1,9 @@
 -- migrate:up
 CREATE TABLE IF NOT EXISTS sales_local ON CLUSTER '{cluster}' (
     WEEK Date32,
-    COUNTRY_ID Decimal(38, 9),
+    COUNTRY_ID UInt64,
     REGION String,
-    PRODUCT_ID Nullable(Decimal(38, 10)),
+    PRODUCT_ID Nullable(UInt64),
     UNITS Nullable(Float64),
     DOLLAR_VOLUME Nullable(Decimal(38, 10))
 ) ENGINE = ReplicatedMergeTree('/clickhouse/{installation}/{cluster}/tables/{shard}/{database}/{table}', '{replica}')
