+++ databases/test/migrations/20220415232010_create_distributed_sales_table.sql
@@ -1,9 +1,9 @@
 -- migrate:up
 CREATE TABLE IF NOT EXISTS sales_distributed ON CLUSTER '{cluster}' (
     WEEK Date32,
-    COUNTRY_ID Decimal(38, 9),
+    COUNTRY_ID UInt64,
     REGION String,
-    PRODUCT_ID Nullable(Decimal(38, 10)),
+    PRODUCT_ID Nullable(UInt64),
     UNITS Nullable(Float64),
     DOLLAR_VOLUME Nullable(Decimal(38, 10))
 ) ENGINE = Distributed('{cluster}', currentDatabase(), sales_local, rand());
