+++ databases/test/migrations/20220711214932_create_distributed_product_dim.sql
@@ -1,6 +1,6 @@
 -- migrate:up
 CREATE TABLE IF NOT EXISTS products_distributed ON CLUSTER '{cluster}' (
-    PRODUCT_ID Decimal(38, 10),
+    PRODUCT_ID UInt64,
     PRODUCT_NAME String
 ) ENGINE = Distributed('{cluster}', currentDatabase(), products_local, rand());
